+++ spirit/topic/models.py
@@ -44,6 +44,15 @@ class Topic(models.Model):
         else:
             return reverse('spirit:topic:detail', kwargs={'pk': str(self.id), 'slug': self.slug})
 
+    def get_bookmark_url(self):
+        if not self.is_visited:
+            return self.get_absolute_url()
+
+        if not self.has_new_comments:
+            return self.bookmark.get_absolute_url()
+
+        return self.bookmark.get_new_comment_url()
+
     @property
     def main_category(self):
         return self.category.parent or self.category
@@ -71,6 +80,10 @@ class Topic(models.Model):
     @property
     def has_new_comments(self):
         return self.new_comments_count > 0
+
+    @property
+    def is_visited(self):
+        return bool(self.bookmark)
 
     def increase_view_count(self):
         Topic.objects\